Update dash.css
Hi Leon! Now that I got access to my email I was able to finish my Github account setup. I copied/pasted some of the code that I wrote. Not done with this yet, but would love to test and see if it works like I'm hoping it will without my styles being overwritten. I was careful to use the page-id-class so my styles shouldn't interfere with other pages. 

Is there a way for me to get access to the stylesheet so that I can make changes to it without having to wait for approval? It's hard to know if my code will work when I can't preview it as I work. I was unable to get the WordPress custom code thing to work (when I tried to publish, it said I didn't have access).